Add files via upload
This code uses PySpark to read a text file from HDFS, count the number of lines and errors in the file, and display some error messages. It uses Spark’s DataFrame API to perform these operations. It also specifies ‘yarn’ as the master, which means it will run on a Hadoop cluster using YARN as the resource manager.